@dinahshi 2021-02-10
Upon closer review, I misread the benchmark numbers! It seems this change is slower than `main`, investigating now. Please consider this PR on pause! 🙇🏻‍♀️