@dinahshi 2021-02-10
Performance benchmark numbers fixed! There was an error in the benchmarking script where the association was not reset after warm up / between runs. Edited main description with the fix to the script.

Now the numbers are more in line with expectations. There is an improvement where queries are grouped as 1 query is executed instead of 2. The reduction would likely be more dramatic using a different database like MySQL where database round trips are more costly. The control case where no queries are grouped is slightly slower as we've added a call to `to_sql`. 

With that, I believe this change is ready to merge. Thank you!